feat(geoutils): tree-shakeable getAdministrativeRegionById function
