+++ src/__tests__/geoUtils.test.ts
@@ -15,7 +15,6 @@ import {
   MOUNT_ATHOS_PREFECTURE_ID,
   MOUNT_ATHOS_REGION_ID,
   findByPostalCode,
-  getAdministrativeRegionById,
   getAdministrativeRegionByIsoCode,
   getAdministrativeUnitById,
   getAdministrativeUnits,
@@ -55,106 +54,6 @@ const prefecturesWithoutMountAthos = {
 };
 
 const allCountries = { el: countriesEl, en: countriesEn } as const;
-
-describe("getAdministrativeRegionById", () => {
-  it("correctly returns region with default values (in greek language)", () => {
-    const expectedData = administrativeRegionsWithoutMountAthos.el[0];
-
-    expect(getAdministrativeRegionById({ id: 1 })).toEqual(expectedData);
-    expect(getAdministrativeRegionById({ id: 1, locale: "el" })).toEqual(expectedData);
-    expect(getAdministrativeRegionById({ id: 1, includeMountAthos: false })).toEqual(expectedData);
-    expect(getAdministrativeRegionById({ id: 1, level: "municipality" })).toEqual(expectedData);
-    // all default options
-    expect(
-      getAdministrativeRegionById({
-        id: 1,
-        locale: "el",
-        includeMountAthos: false,
-        level: "municipality",
-      }),
-    ).toEqual(expectedData);
-  });
-
-  it("correctly returns Mount Athos region (in greek language)", () => {
-    const expectedData = administrativeRegions.el[13];
-
-    expect(getAdministrativeRegionById({ id: 14, includeMountAthos: true })).toEqual(expectedData);
-    expect(getAdministrativeRegionById({ id: 14, locale: "el", includeMountAthos: true })).toEqual(expectedData);
-    expect(
-      getAdministrativeRegionById({
-        id: 14,
-        locale: "el",
-        includeMountAthos: true,
-        level: "municipality",
-      }),
-    ).toEqual(expectedData);
-  });
-
-  it("correctly returns region data with correct level (in greek language)", () => {
-    const expectedRegionLevelData = administrativeRegionsWithoutMountAthos.el.map(
-      ({ units: _units, ...region }) => region,
-    )[7];
-    const expectedUnitLevelData = administrativeRegionsWithoutMountAthos.el.map((region) => ({
-      ...region,
-      units: region.units.map(({ municipalities: _municipalities, ...unit }) => unit),
-    }))[7];
-    const expectedMunicipalityLevelData = administrativeRegionsWithoutMountAthos.el[7];
-
-    expect(getAdministrativeRegionById({ id: 8, level: "region" })).toStrictEqual(expectedRegionLevelData);
-    expect(getAdministrativeRegionById({ id: 8, level: "unit" })).toStrictEqual(expectedUnitLevelData);
-    expect(getAdministrativeRegionById({ id: 8, level: "municipality" })).toStrictEqual(expectedMunicipalityLevelData);
-  });
-
-  it("correctly returns region (in english language)", () => {
-    const expectedData = administrativeRegionsWithoutMountAthos.en[3];
-
-    expect(getAdministrativeRegionById({ id: 4, locale: "en" })).toEqual(expectedData);
-    expect(getAdministrativeRegionById({ id: 4, locale: "en", includeMountAthos: false })).toEqual(expectedData);
-    expect(getAdministrativeRegionById({ id: 4, locale: "en", level: "municipality" })).toEqual(expectedData);
-    // all default options
-    expect(
-      getAdministrativeRegionById({
-        id: 4,
-        locale: "en",
-        includeMountAthos: false,
-        level: "municipality",
-      }),
-    ).toEqual(expectedData);
-  });
-
-  it("correctly returns Mount Athos region (in english language)", () => {
-    const expectedData = administrativeRegions.en[13];
-
-    expect(getAdministrativeRegionById({ id: 14, locale: "en", includeMountAthos: true })).toEqual(expectedData);
-    expect(
-      getAdministrativeRegionById({
-        id: 14,
-        locale: "en",
-        includeMountAthos: true,
-        level: "municipality",
-      }),
-    ).toEqual(expectedData);
-  });
-
-  it("correctly returns region data with correct level (in english language)", () => {
-    const expectedRegionLevelData = administrativeRegionsWithoutMountAthos.en.map(
-      ({ units: _units, ...region }) => region,
-    )[5];
-    const expectedUnitLevelData = administrativeRegionsWithoutMountAthos.en.map((region) => ({
-      ...region,
-      units: region.units.map(({ municipalities: _municipalities, ...unit }) => unit),
-    }))[5];
-    const expectedMunicipalityLevelData = administrativeRegionsWithoutMountAthos.en[5];
-
-    expect(getAdministrativeRegionById({ id: 6, locale: "en", level: "region" })).toStrictEqual(
-      expectedRegionLevelData,
-    );
-    expect(getAdministrativeRegionById({ id: 6, locale: "en", level: "unit" })).toStrictEqual(expectedUnitLevelData);
-    expect(getAdministrativeRegionById({ id: 6, locale: "en", level: "municipality" })).toStrictEqual(
-      expectedMunicipalityLevelData,
-    );
-  });
-});
 
 describe("getAdministrativeRegionByIsoCode", () => {
   it("correctly returns region with default values (in greek language)", () => {
